refactor: split api_server on module-level
