+++ gpt2giga/app/cli.py
@@ -1,0 +1,24 @@
+"""CLI-driven application configuration loading."""
+
+import argparse
+import os
+
+from dotenv import find_dotenv, load_dotenv
+
+from gpt2giga.common.app_meta import warn_sensitive_cli_args
+from gpt2giga.models.config import ProxyConfig
+
+
+def load_config() -> ProxyConfig:
+    """Load configuration from CLI arguments and environment variables."""
+    warn_sensitive_cli_args()
+
+    parser = argparse.ArgumentParser(add_help=False)
+    parser.add_argument("--env-path", type=str, default=None, help="Path to .env file")
+    args, _ = parser.parse_known_args()
+
+    requested_env = args.env_path if args.env_path else f"{os.getcwd()}/.env"
+    env_path = find_dotenv(requested_env)
+    load_dotenv(env_path)
+
+    return ProxyConfig()
